fix: guard JSON.parse in useCollection against malformed filters param
Wraps the JSON.parse call for the `filters` URL param in a try/catch so
a user-crafted URL like ?filters=hello falls back to an empty array
instead of crashing the CollectionClient subtree with a SyntaxError.

Also adds @testing-library/react to packages/components and migrates
useCollection and useQueryParams tests from manual createRoot/HookConsumer
boilerplate to renderHook, including new tests covering the fix.
